Type revision (cont.): complete more dependencies to the new type system

+++ spark/src/main/scala/io/ddf/spark/content/DataFrame2MatrixVector.scala
@@ -105,12 +105,12 @@ class DataFrame2MatrixVector(@transient ddf: DDF) extends ConvertFunction(ddf) {
           (row: Row, idx: Int) => row.getInt(idx).toDouble
         }
         case ColumnType.DOUBLE => {
-          (row: Row, idx: Int) => row.getDouble(idx).toDouble
+          (row: Row, idx: Int) => row.getDouble(idx)
         }
         case ColumnType.FLOAT => {
           (row: Row, idx: Int) => row.getFloat(idx).toDouble
         }
-//        getDecimal is available only in Spark 1.3.1
+//        //getDecimal is available only in Spark 1.3.1
 //        case ColumnType.DECIMAL => {
 //          (row: Row, idx: Int) => row.getDecimal(idx).toDouble
 //        }
